Modify the naming of the compilation container
When the compilation container is created for the first time, the code
file directory under the current directory A is mounted to /repo.

In the following case, when the user switches the project to another
directory, for example, directory B, if the user runs oneshot.py again,
the code files in directory B will not be compiled, and the user will be
confused about the result if he/she does not look at the py file carefully.